Enable more tests
* Enable test_workers.py renamed to test_workflows.py
* Rename test_workflows.py to test_cli.py
