+++ oonipipeline/tests/data/.gitignore
@@ -1,2 +1,3 @@
 /datadir
 /measurements
+/raw_measurements
